Properly indented and formatted the code

+++ src/main/java/com/mountblue/blogapplication/controllers/CommentController.java
@@ -1,7 +1,9 @@
 package com.mountblue.blogapplication.controllers;
 
+import com.mountblue.blogapplication.configurations.LoginUser;
 import com.mountblue.blogapplication.entities.Comment;
 import com.mountblue.blogapplication.entities.Post;
+import com.mountblue.blogapplication.entities.User;
 import com.mountblue.blogapplication.services.CommentService;
 import com.mountblue.blogapplication.services.PostService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -20,10 +22,14 @@ public class CommentController {
 
     private final CommentService commentService;
     private  final PostService postService;
+
+    private  final LoginUser loginUser;
+
     @Autowired
-    public CommentController(CommentService commentService, PostService postService) {
+    public CommentController(CommentService commentService, PostService postService, LoginUser loginUser) {
         this.commentService = commentService;
         this.postService = postService;
+        this.loginUser = loginUser;
     }
 
     @PostMapping("/add-comment/{postId}")
@@ -42,11 +48,19 @@ public class CommentController {
     }
 
     @GetMapping("/edit-comment/{id}")
-    public  String showEditCommentPage(@PathVariable Integer id, Model model){
+    public  String editComment(@PathVariable Integer id, Model model){
         Comment comment=commentService.findCommentById(id);
-        model.addAttribute("post",comment.getPost());
-        model.addAttribute("commentId",id);
-        return "editcomment";
+
+        Post post = comment.getPost();
+        User currentUser= loginUser.getDetails();
+
+        if(currentUser.getRole().equals("ADMIN")||currentUser.getPosts().contains(post)) {
+            model.addAttribute("post", comment.getPost());
+            model.addAttribute("commentId", id);
+            return "edit-comment";
+        }
+
+        return "redirect:/post/"+post.getId();
     }
 
     @PostMapping("/update-comment")
@@ -58,9 +72,15 @@ public class CommentController {
     }
 
     @GetMapping("/delete-comment/{id}")
-    public String deleteComment(@PathVariable Integer id){
-        Comment comment=commentService.findCommentById(id);
-        commentService.deleteCommentById(id);
-        return "redirect:/post/"+comment.getPost().getId();
+    public String deleteComment(@PathVariable Integer id) {
+
+        Comment comment = commentService.findCommentById(id);
+        Post post = comment.getPost();
+        User currentUser = loginUser.getDetails();
+        if (currentUser.getRole().equals("ADMIN") || currentUser.getPosts().contains(post)) {
+            commentService.deleteCommentById(id);
+
+        }
+        return "redirect:/post/" + post.getId();
     }
 }
